% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nba-per-game-stats.R
\name{GetNBAPerGameStatistics}
\alias{GetNBAPerGameStatistics}
\title{This function returns all player statistics from an NBA season on
basketball-reference.com.  See an example table at:
http://www.basketball-reference.com/leagues/NBA_2015_per_game.html}
\usage{
GetNBAPerGameStatistics(season = 2016)
}
\arguments{
\item{season}{A numeric year}
}
\value{
An object of class tbl_df
}
\description{
This function returns all player statistics from an NBA season on
basketball-reference.com.  See an example table at:
http://www.basketball-reference.com/leagues/NBA_2015_per_game.html
}
\examples{
GetNBAPerGameStatistics(season = 2015)
players <- GetNBAPerGameStatistics(season = 2015) \%>\%
  filter(MP > 20, Pos \%in\% c("SF")) \%>\%
  select(Player, link) \%>\%
  distinct()
}

