% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nba-player-per-game-stats.R
\name{GetNBAPlayerPerGameStats}
\alias{GetNBAPlayerPerGameStats}
\title{NBA Player Career Statistics}
\usage{
GetNBAPlayerPerGameStats(player_link)
}
\arguments{
\item{player_link}{A link suffix, e.g. "/players/d/davisan02.html"}
}
\value{
An object of class tbl_df
}
\description{
This function gets a player's career stats from basketball-reference.com
}
\examples{
GetNBAPlayerPerGameStats("/players/d/davisan02.html") # Anthony Davis
GetNBAPlayerPerGameStats("/players/j/jamesle01.html") # Lebron James
}

