% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-player-per-game-stats.R
\name{NBAPlayerPerGameStats}
\alias{NBAPlayerPerGameStats}
\title{NBA Player Career Statistics}
\usage{
NBAPlayerPerGameStats(player_link)
}
\arguments{
\item{player_link}{A link suffix, e.g. "/players/d/davisan02.html"}
}
\value{
An object of class tbl_df
}
\description{
This function gets a player's career stats from basketball-reference.com
}
\examples{
NBAPlayerPerGameStats("/players/d/davisan02.html") # Anthony Davis
NBAPlayerPerGameStats("/players/j/jamesle01.html") # Lebron James
}
