% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-season-team-by-year.R
\name{NBASeasonTeamByYear}
\alias{NBASeasonTeamByYear}
\title{NBA Team Statistics By Year}
\usage{
NBASeasonTeamByYear(team, season)
}
\arguments{
\item{team}{One of "ATL", "BOS", etc.}

\item{season}{The particular season you are querying}
}
\value{
A team's game statistics for a particular season
}
\description{
This function returns a team's yearly statistics from basketball-reference.com
}
\examples{
NBASeasonTeamByYear("ATL", 2015)
NBASeasonTeamByYear("NOP", 2015)
}
