% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nba-standings-by-date.R
\name{NBAStandingsByDate}
\alias{NBAStandingsByDate}
\title{NBA Standings by Date}
\usage{
NBAStandingsByDate(date_string = Sys.Date())
}
\arguments{
\item{date_string}{A String of the form "2015-04-01"}
}
\value{
An list containing the standings in the Eastern and Western Conferences
}
\description{
This function the current NBA standings for the given date
}
\examples{
NBAStandingsByDate("2010-01-31") # Jan 31, 2010
NBAStandingsByDate("2015-11-19") # Nov 19, 2015
}

