% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-standings-by-date.R
\name{NBAStandingsByDate}
\alias{NBAStandingsByDate}
\title{NBA Standings by Date}
\usage{
NBAStandingsByDate(date_string = Sys.Date())
}
\arguments{
\item{date_string}{A String of the form "2015-04-01"}
}
\value{
An list containing the standings in the Eastern and Western Conferences
}
\description{
This function the current NBA standings for the given date
}
\examples{
NBAStandingsByDate("2010-01-31") # Jan 31, 2010
NBAStandingsByDate("2017-11-09") # Nov 09, 2017
}
